Added authentication in every controller. Added friendship method

+++ CleanArch.API/Controllers/MessageController.cs
@@ -1,10 +1,13 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
 using RPGOnline.Application.DTOs.Requests;
 using RPGOnline.Application.DTOs.Requests.Mail;
 using RPGOnline.Application.Interfaces;
+using System.Security.Claims;
 
 namespace RPGOnline.API.Controllers
 {
+    [Authorize]
     public class MessageController : CommonController
     {
 
@@ -22,6 +25,11 @@ namespace RPGOnline.API.Controllers
         {
             try
             {
+                if (!IsSameId(id))
+                {
+                    return BadRequest("Access denied - bad ID");
+                }
+
                 var result = await _messageService.GetUserMessages(id, getMessageRequest);
                 return Ok(new
                 {
@@ -40,9 +48,12 @@ namespace RPGOnline.API.Controllers
         [HttpPost("{senderId}/send")]
         public async Task<IActionResult> PostMessage(int senderId, PostMessageRequest messageRequest)
         {
-
             try
             {
+                if (!IsSameId(senderId))
+                {
+                    return BadRequest("Access denied - bad ID");
+                }
 
                 var result = await _messageService.PostMessage(senderId, messageRequest);
 
@@ -65,6 +76,11 @@ namespace RPGOnline.API.Controllers
         {
             try
             {
+                if (!IsSameId(uId))
+                {
+                    return BadRequest("Access denied - bad ID");
+                }
+
                 var result = await _messageService.DeleteMessage(uId, messageId);
 
                 if (result == null)
@@ -77,6 +93,14 @@ namespace RPGOnline.API.Controllers
             {
                 return BadRequest(ex.Message);
             }
+        }
+
+        private bool IsSameId(int id)
+        {
+            var claimsIdentity = this.User.Identity as ClaimsIdentity;
+            var userId = claimsIdentity.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+
+            return userId.Equals(id.ToString());
         }
     }
 }
